+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.18.0] - 2020-04-15
+### Added
+- Add new address function named `setDeliveryAddressOnList`
+
 ## [2.17.2] - 2020-03-18
 ### Added
 - Options to number verification at `isAddressComplete` function